[unittest] Moved updateAllPinocchio to unittest code
