+++ include/crocoddyl/multibody/numdiff/cost.hpp
@@ -13,37 +13,8 @@
 #include <boost/function.hpp>
 #include "crocoddyl/multibody/fwd.hpp"
 #include "crocoddyl/multibody/cost-base.hpp"
-#include "pinocchio/algorithm/kinematics.hpp"
-#include "pinocchio/algorithm/jacobian.hpp"
-#include "pinocchio/algorithm/frames.hpp"
-#include "pinocchio/algorithm/center-of-mass.hpp"
-#include "pinocchio/algorithm/centroidal.hpp"
-#include "pinocchio/algorithm/centroidal-derivatives.hpp"
 
 namespace crocoddyl {
-
-/**
- * @brief Compute all the pinocchio data needed for the numerical
- * differentiation. We use the address of the object to avoid a copy from the
- * "boost::bind".
- *
- * @param model is the rigid body robot model.
- * @param data contains the results of the computations.
- * @param x is the state vector.
- */
-void updateAllPinocchio(pinocchio::Model* const model, pinocchio::Data* data, const Eigen::VectorXd& x) {
-  const Eigen::VectorXd& q = x.segment(0, model->nq);
-  const Eigen::VectorXd& v = x.segment(model->nq, model->nv);
-  Eigen::VectorXd a = Eigen::VectorXd::Zero(model->nv);
-  Eigen::Matrix<double, 6, Eigen::Dynamic> tmp;
-  tmp.resize(6, model->nv);
-  pinocchio::forwardKinematics(*model, *data, q);
-  pinocchio::computeJointJacobians(*model, *data, q);
-  pinocchio::updateFramePlacements(*model, *data);
-  pinocchio::jacobianCenterOfMass(*model, *data, q);
-  pinocchio::computeCentroidalMomentum(*model, *data, q, v);
-  pinocchio::computeCentroidalDynamicsDerivatives(*model, *data, q, v, a, tmp, tmp, tmp, tmp);
-}
 
 template <typename _Scalar>
 class CostModelNumDiffTpl : public CostModelAbstractTpl<_Scalar> {
