+++ unittest/test_costs.cpp
@@ -2,7 +2,7 @@
 // BSD 3-Clause License
 //
 // Copyright (C) 2018-2020, LAAS-CNRS, New York University, Max Planck Gesellschaft
-//                          University of Edinburgh, INRIA
+//                          University of Edinburgh, INRIA,
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -101,11 +101,11 @@ void test_partial_derivatives_against_numdiff(CostModelTypes::Type cost_type,
   const Eigen::VectorXd& u = Eigen::VectorXd::Random(model->get_nu());
 
   // Compute all the pinocchio function needed for the models.
-  crocoddyl::updateAllPinocchio(&pinocchio_model, &pinocchio_data, x);
+  crocoddyl_unit_test::updateAllPinocchio(&pinocchio_model, &pinocchio_data, x);
 
   // set the function that needs to be called at every step of the numdiff
   std::vector<crocoddyl::CostModelNumDiff::ReevaluationFunction> reevals;
-  reevals.push_back(boost::bind(&crocoddyl::updateAllPinocchio, &pinocchio_model, &pinocchio_data, _1));
+  reevals.push_back(boost::bind(&crocoddyl_unit_test::updateAllPinocchio, &pinocchio_model, &pinocchio_data, _1));
   model_num_diff.set_reevals(reevals);
 
   // Computing the action derivatives
